requirements.txt
langchain==0.3.7
langchain_openai==0.2.8
pytest==8.3.3
python-dotenv==1.0.1
streamlit==1.40.0
